creo que funciona todo

+++ src/deleteListing.js
@@ -1,3 +1,4 @@
+import * as admin from 'firebase-admin'
 import { db } from "./database";
 
 export const deleteListingRoute = {
@@ -6,11 +7,12 @@ export const deleteListingRoute = {
     path: '/api/listings/{id}',
     handler: async (req, h)=>{
         const {id} = req.params;
-        // Later on we will have to validate the user that is making the request
-
+        const token = req.headers.authtoken;
+        const user = await admin.auth().verifyIdToken(token); 
+        const userId = user.user_id;
         await db.query(
-            'DELETE FROM listings WHERE id=?',
-            [id], 
+            'DELETE FROM listings WHERE id=? AND user_id=?',
+            [id, userId], 
         );
         return {message: 'The register has been deleted, hooray!'}; 
     }
